Add game structure with main menu, leaderboard, and player controls into classes
