+++ src/Entity/Player.py
@@ -1,37 +1,14 @@
+from src.Entity.Entity import Entity
+from src.constant import WIDTH, vec, ACC, FRIC, platforms
+from pygame import *
 import pygame
-import sys
-from pygame.locals import *
-import time
 import os
-
-# Initialize Pygame
-pygame.init()
-
-# Constants
-WIDTH = 800
-HEIGHT = 600
-FPS = 60
-VEC = pygame.math.Vector2
-ACC = 0.5
-FRIC = -0.12
-vec = pygame.math.Vector2
-
-# Setup display
-displaysurface = pygame.display.set_mode((WIDTH, HEIGHT))
-pygame.display.set_caption("Game")
-FramePerSec = pygame.time.Clock()
-
-# Font setup
-font = pygame.font.SysFont("Arial", 20)
-
-# Global variables for access in other modules
-platforms = pygame.sprite.Group()
-all_sprites = pygame.sprite.Group()
+import time
 
 
-class Player(pygame.sprite.Sprite):
-    def __init__(self):
-        super().__init__()
+class Player(Entity):
+    def __init__(self, width=120, height=120, x=10, y=385):
+        super().__init__(pos=(x, y), size=(width, height), color=(128, 255, 40))
 
         # Animation variables
         self.animation_frames = []
@@ -43,25 +20,16 @@ class Player(pygame.sprite.Sprite):
         self.static_image = None
         self.moving = False
         self.dashing = False
+        self.jumping = False
 
-        # Load static image and animation frames
+        # Load images
         self.load_images()
 
-        # Set initial surface
+        # Override initial surface if images are loaded
         if self.static_image:
             self.surf = self.static_image
         elif self.animation_frames:
             self.surf = self.animation_frames[0]
-        else:
-            # Fallback to a colored rectangle
-            self.surf = pygame.Surface((30, 30))
-            self.surf.fill((128, 255, 40))
-
-        self.rect = self.surf.get_rect()
-        self.pos = vec((10, 385))
-        self.vel = vec(0, 0)
-        self.acc = vec(0, 0)
-        self.jumping = False
 
     def load_images(self):
         try:
@@ -201,69 +169,3 @@ class Player(pygame.sprite.Sprite):
                 self.pos.y = hits[0].rect.top
                 self.vel.y = 0
                 self.jumping = False
-
-
-class Platform(pygame.sprite.Sprite):
-    def __init__(self):
-        super().__init__()
-        self.surf = pygame.Surface((WIDTH, 20))
-        self.surf.fill((255, 0, 0))
-        self.rect = self.surf.get_rect(center=(WIDTH / 2, HEIGHT - 10))
-
-
-def initialize_game():
-    global platforms, all_sprites
-
-    # Clear previous sprites if any
-    platforms.empty()
-    all_sprites.empty()
-
-    # Create new game objects
-    PT1 = Platform()
-    P1 = Player()
-
-    # Add them to the groups
-    platforms.add(PT1)
-    all_sprites.add(PT1)
-    all_sprites.add(P1)
-
-    return P1, PT1, platforms, all_sprites
-
-
-def run_game(P1, all_sprites):
-    """Run the main game loop without menu system"""
-    while True:
-        for event in pygame.event.get():
-            if event.type == QUIT:
-                pygame.quit()
-                sys.exit()
-            elif event.type == KEYDOWN:
-                if event.key == K_ESCAPE:
-                    pygame.quit()
-                    sys.exit()
-
-        displaysurface.fill((0, 0, 0))
-
-        P1.move()
-        P1.update()
-        for entity in all_sprites:
-            displaysurface.blit(entity.surf, entity.rect)
-
-        # Display FPS
-        fps = int(FramePerSec.get_fps())
-        fps_text = font.render(f"FPS: {fps}", True, (255, 255, 255))
-        displaysurface.blit(fps_text, (10, 10))
-
-        # Display player coordinates
-        pos_text = font.render(
-            f"X: {int(P1.pos.x)}, Y: {int(P1.pos.y)}", True, (255, 255, 255)
-        )
-        displaysurface.blit(pos_text, (10, 40))
-
-        pygame.display.update()
-        FramePerSec.tick(FPS)
-
-
-if __name__ == "__main__":
-    P1, PT1, platforms, all_sprites = initialize_game()
-    run_game(P1, all_sprites)
